+deps for yyjson; making layout for build.sh more consistent

+++ pcre2/build.sh
@@ -1,21 +1,28 @@
 #!/bin/bash
 set -e
-depends=()
-lfs_depends=(bash bzip2coreutils glibc gzip make readline sed tar)
-blfs_depends=(wget)
+# Variable declarations
 name=pcre2
 version=$(wget -cqO- https://github.com/PCRE2Project/pcre2/releases | grep "releases/tag/pcre2" | grep -v "alpha\|beta\|rc" | head -n 1 | cut -d '"' -f 6 | cut -d '=' -f 3 | cut -d '/' -f 6 | sed 's/pcre2-//g')
 filename="$name-$version.tar.gz"
-wget -c https://github.com/PCRE2Project/pcre2/archive/$filename
-wget -c https://github.com/zherczeg/sljit/archive/master.tar.gz -O sljit-master.tar.gz
-tar xf $name-$version.tar.gz
-dir_name="$name-$name-$version"
+direname="$name-$name-$version"
+depends=()
+lfs_depends=(bash bzip2 coreutils glibc gzip make readline sed tar)
+blfs_depends=(wget)
+# Fetch and unpack source
+rm -rf $direname
+if ! [[ -f $filename ]]; then
+	wget -c https://github.com/PCRE2Project/pcre2/archive/$filename
+fi
+if ! [[ -f sljit-master.tar.gz ]]; then
+	wget -c https://github.com/zherczeg/sljit/archive/master.tar.gz -O sljit-master.tar.gz
+fi
+tar xf $filename
 cd $direname
 rm -rf deps/sljit
 tar xf ../sljit-master.tar.gz
 mv sljit-master sljit 
 mv sljit deps/sljit
-
+# Compile and install
 ./autogen.sh
 configure_options=(
     --enable-jit
@@ -26,13 +33,12 @@ configure_options=(
     --enable-pcre2test-libreadline
     --prefix=/usr
   )
-
 CFLAGS+="-O2 -fPIC -ffat-lto-objects"
 CXXFLAGS+="-O2 -fPIC -ffat-lto-objects"
-
 ./configure "${configure_options[@]}"
 make -j$(nproc)
 sudo make install
+# Cleanup and add to database
 cd ..
 sudo rm -rf ${filename} $direname
 echo $version > /var/lib/lfs-custom-packages/$name
