+++ qrupdate/build.sh
@@ -1,17 +1,22 @@
 #!/bin/bash
 set -e
+# Variable declarations
+name=qrupdate
+version=$(wget -cqO- https://github.com/mpimd-csc/qrupdate-ng/releases | grep "releases/tag/v" | grep -v "alpha\|beta\|rc" | cut -d '"' -f 6 | cut -d '/' -f 6 | sed 's/^v//g')
+filename=$name-$version.tar.gz
+direname="$name-ng-$version"
 depends=(blas lapack)
 lfs_depends=(bash coreutils glibc gzip make sed tar)
 blfs_depends=(cmake
 gcc # Fortran support needed
 wget)
-name=qrupdate
-version=$(wget -cqO- https://github.com/mpimd-csc/qrupdate-ng/releases | grep "releases/tag/v" | grep -v "alpha\|beta\|rc" | cut -d '"' -f 6 | cut -d '/' -f 6 | sed 's/^v//g')
-filename=$name-$version.tar.gz
-direname="$name-ng-$version"
-wget -c https://github.com/mpimd-csc/qrupdate-ng/archive/v$version.tar.gz -O $filename
+# Fetch and unpack source
+if ! [[ -f $filename ]]; then
+	wget -c https://github.com/mpimd-csc/qrupdate-ng/archive/v$version.tar.gz -O $filename
+fi
 rm -rf $direname
 tar xf $filename
+# Compile and install
 cd $direname
 CFLAGS="-O2 -fPIC"
 CXXFLAGS="-O2 -fPIC"
@@ -22,6 +27,7 @@ cmake -B build -S . \
     -DCMAKE_CXX_FLAGS="$CXXFLAGS"
 cmake --build build --verbose
 sudo cmake --install build
+# Cleanup and add to database
 cd ..
 sudo rm -rf $filename $direname
 echo $version > /var/lib/lfs-custom-packages/$name
